src/requirements.txt
requests==2.27.1
polling2==0.5.0
PyYAML==6.0
semver==2.13.0
cryptography==39.0.0
kubernetes==25.3.0
kubeconfig-1.1.1
